Support reindexing updated documents
I decided to support indexing arbitrary new documents under a given
ID. To make it more like an external index, I also chose that we
couldn't make the assumption we had access to the document we previously
indexed.

The first step of updating a document in the index is removing the
previously indexed values for the document. To find out what the
previously indexed values are, because we don't have access to the
old version of the document, we have to create a forward index of
docID -> key, so we're able to look up the keys we previously indexed
for the against that ID. This feel inefficient, but I couldn't think
of a better way. We actually pack all the information into the pebble
key, because that was an easy way to store many keys against one ID.

After we've removed the old document's values, we can index as usual.
An optimisation I didn't do was to generate the pathValueKeys for the
new document, and compare them to the old ones, so we could update
the delta rather than remove and add uncessarily.