Make searchIndex work
searchIndex takes a parsed query and executes it on the index. It
supports =, <, > query types, which map to index operations lookupEq,
lookupLT, lookupGT.

This commit adds tests for searchIndex and fixes it to work with all
the index types we have (eg, fix use of lookupGE to lookupGT).

We also allow for query types to pass in any type of supported value,
in addition to strings that is booleans, numbers and nil.

The two methods in main that deal with querying are commented out
because they don't need to be updated to work with the new code
yet.

searchIndex requires memory proportionate to the number of ids
returned by all the index queries that it executes, which is more
than strictly needed but okay for now.