feat: embed doc ID in index key
Instead of storing the doc ID in a postings list for each field, value
combination, put the doc ID directly into the key. So change from:

[invIdxNS, path value] = doc1, doc2

to

[invIdxNS, path, value, doc1]
[invIdxNS, path, value, doc2]

This has the nice property that iterating the results of a key-value
store query returns the doc IDs in order for each field, value
combination, although we don't use that fact. I think that, even without
this, the change does simplify the code in various places, particularly
by removing the code that manages the postings list.

We do have to get smarter about decoding the inverted index key,
however. As the value may contain 00 bytes (particularly numbers due
to our encoding), we can't just split on 00 bytes (unpackTuple) any
more. Instead, we add a custom decode function, decodeInvIdxKey. We
assume this is exercised fine by the Test_lookupXXX tests.

Overall, it feels like we shift a problem into the key-value store,
that of managing the docIDs, without meaningfully changing the
complexity of our code.